allow for empty outputs of agfusion to pass through the final steps

+++ modules/local/agfusion/batch/main.nf
@@ -14,8 +14,7 @@ process AGFUSION_BATCH {
     path(pyensembl_cache)
 
     output:
-    tuple val(meta), path("${prefix}/**")                    , emit: fusions_annotated
-    tuple val(meta), path("${prefix}.fusion_transcripts.csv"), emit: fusion_transcripts_csv
+    tuple val(meta), path("${prefix}/**")                    , emit: fusions_annotated, optional: true
     tuple val(meta), path("${prefix}.fusion_transcripts.tsv"), emit: fusion_transcripts_tsv
     path "versions.yml"                                      , emit: versions
 
@@ -29,15 +28,19 @@ process AGFUSION_BATCH {
     export PYENSEMBL_CACHE_DIR=\$PWD/${pyensembl_cache}
 
     awk -F"\\t" 'NR == 1 && \$1 ~ /gene5_chr/ {next;}{print}' ${fusions} > ${fusions}.no_header
+    
+    if [ -s ${fusions}.no_header ]; then
+        agfusion batch \\
+            -f ${fusions}.no_header \\
+            -db ${agfusion_db} \\
+            -o ${prefix} \\
+            ${args}
 
-    agfusion batch \\
-        -f ${fusions}.no_header \\
-        -db ${agfusion_db} \\
-        -o ${prefix} \\
-        ${args}
-
-    awk -F"," 'NR != 1 && FNR == 1 {next;}{print}' ${prefix}/*/*.fusion_transcripts.csv > ${prefix}.fusion_transcripts.csv
-    cat ${prefix}.fusion_transcripts.csv | tr "," "\\t" > ${prefix}.fusion_transcripts.tsv
+        awk -F"," 'NR != 1 && FNR == 1 {next;}{print}' ${prefix}/*/*.fusion_transcripts.csv > ${prefix}.fusion_transcripts.csv
+        cat ${prefix}.fusion_transcripts.csv | tr "," "\\t" > ${prefix}.fusion_transcripts.tsv
+    else 
+        echo -e "3'_transcript\t5'_transcript\t5'_gene\t3'_gene\t5'_breakpoint\t3'_breakpoint\t5'_strand\t3'_strand\t5'_transcript_biotype\t3'_transcript_biotype" > ${prefix}.fusion_transcripts.tsv
+    fi
 
     cat <<-END_VERSIONS > versions.yml
     "${task.process}":
